필요없는 씬들, 타일들 없앴고, 10초 돌다가 펑 터지는 새로운 버그 수정. 이제 몬스터 스포너로 몬스터 스폰하게 하자.

+++ yaExpGemObjPool.cpp
@@ -41,10 +41,10 @@ namespace ya
 	}
 	UINT ExpGemObjPool::GetFreeObjectCount() const
 	{
-		return mQ.size();
+		return static_cast<UINT>(mQ.size());
 	}
 	UINT ExpGemObjPool::GetMaxFreeObjectCount() const
 	{
-		return mMaxPoolSize;
+		return static_cast<UINT>(mMaxPoolSize);
 	}
 }
